Update requirements to not include notebook only dependencies
This commit updates the requirements list to only include required
packages for using qrao as a library. The requirements list get
installed unconditionally as part of installing the library meaning that
every user who tries to use qrao will get them installed too. Having
heavy and closed source dependencies like cplex in the list is
problematic for library users in particular because it makes it
difficult for any downstream consumers to depend on the package since it
will always pull those in. To avoid this issue this commit moves the
tutorial specific requirements into an optional extra requirement. This
then separates the concerns of library users and tutorial users. If you
run tutorials you can just run 'pip install prototyp-qrao[tutorials]'
which will install the library with the tutorials extras. At the same
time these dependencies are added to the requirements-dev.txt list too
so that tutorial dependencies are installed during documentation builds
too.