Rename the `tutorial` extra requirement to `notebook-dependencies`
I wanted to change this label, since we will soon have a how-to
guide that depends on cplex (#28), in addition to the existing
tutorial.

I first considered calling it `notebooks`, but it doesn't actually
install any notebooks.  I then settled on `notebook-dependencies`,
which is a bit longer but (I feel) is more clear.