Add retworkx as a direct dependency (#47)
* Add retworkx as a direct dependency

We were previously importing retworkx from `encoding.py` without
specifying it as a dependency, under the assumption that it would
forever be a transitive dependency through qiskit-terra.  However,
qiskit-terra has migrated to the package's new name, rustworkx,
so under this arrangement the latest retworkx package, which is
a backwards compatibility shim, is not installed.  This fixes the
most immediate issue by adding retworkx as a direct dependency

* Depend on the min version of retworkx following min terra version