Use primitives for solving the relaxed problem (#52)
* WIP: use primitives for solving the relaxed problem

* Use QAOAAnsatz with VQE

The new QAOA implementation in qiskit-terra assumes the problem
Hamiltonian is diagonal so that it can benefit from certain
additional features, like applying CVaR or other aggregation
functions.  However, for us the problem Hamiltonian is not
diagonal.  So, we implement QAOA using the `QAOAAnsatz` together
with `VQE`.

* Use the default mixer in QAOA, rather than the problem Hamiltonian

* The optimizer now works with both the old and new eigen[ ]solvers

* Update optimizer to accept new eigensolvers

Parts of this follow https://github.com/Qiskit/qiskit-optimization/pull/436

* Restore tests for the other legacy eigen_solvers

(and suppress the `PendingDeprecationWarning`s using pytest)

* Make the code path for `LegacyNumPyMinimumEigensolver` more explicit

This way, it will be more clear when this branch/block can be
removed in the future

* Fix a typo